proper handling of dev databases

+++ scripts/create-users.js
@@ -1,15 +1,21 @@
-## run with: $ mongo < create-users.mdb
+/*
+ * usage: mongo --eval "let dbname='SB_NAME'" create-users.js
+ */
 
-## find the database
-db = db.getSiblingDB('spectre');
+// check dbname arg
+if (typeof dbname === 'undefined' ||  (dbname !== 'spectre' && dbname !== 'spectre-dev' )) 
+  throw Error("Expected dbname to be= 'spectre' or 'spectre-dev'");
 
-## delete any existing users
+// find the database
+db = db.getSiblingDB(dbname);
+
+// delete any existing users
 db.users.deleteMany({});
 
-## create unique indexes
+// create unique indexes
 db.users.createIndex({ 'login': 1, 'loginType': 1}, { unique: true });
 
-## insert initial users
+// insert initial users
 db.users.insert({
   name: "Daniel",
   login: "daniel@aol.com",
@@ -52,6 +58,5 @@ db.users.insert({
   }
 });
 
-## print all the users name/login/logintype/id
 let cursor = db.users.find({},{login:1,loginType:1,name:1,id:1});
 while (cursor.hasNext()) { printjson(cursor.next()); }
